requirements.txt
streamlit
python-docx
lxml
Pillow